feat: extended functionality of custom alert component

+++ features/baseSlice.js
@@ -1,10 +1,14 @@
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
 
 const getBaseUrl = () => {
-  if (typeof window !== "undefined" && window._env_ && window._env_.API_URL) {
-    return window._env_.API_URL;
+  if (
+    typeof window !== "undefined" &&
+    window._env_ &&
+    window._env_.NEXT_PUBLIC_BASE_URL
+  ) {
+    return window._env_.NEXT_PUBLIC_BASE_URL;
   }
-  return "https:web.builtaccounting.com/"; // Fallback URL if `window` is not defined
+  return process.env.NEXT_PUBLIC_BASE_URL; // Fallback URL if `window` is not defined
 };
 
 const prepareHeaders = (headers, api) => {
@@ -25,7 +29,7 @@ const baseQuery = fetchBaseQuery({
 
 const baseQueryWithReauth = async (args, api, extraOptions) => {
   const result = await baseQuery(args, api, extraOptions);
-  if (result.error && result.error.status === 401) {
+  if (result.error && result.error.status === 401 && api.endpoint !== "login") {
     // log user out
     if (typeof window !== "undefined") {
       localStorage.clear();
